# gh-120145: Speed up symlink lookup for `posixpath.realpath()`