--- a/Misc/NEWS.d/next/Core and Builtins/2024-06-06-05-48-07.gh-issue-120145.1wbqVC.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2024-06-06-05-48-07.gh-issue-120145.1wbqVC.rst
@@ -1,0 +1,1 @@
+Speed up symlink lookup for :func:`os.path.realpath` on Unix.