### Benchmark

<details><summary>Show script</summary>

```sh
# speedup-symlink-lookup.sh
mkdir -p "tmp/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/\
a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a/a"
echo 2 dirs
main/python.exe                   -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/a' * 2"   "os.path.realpath(path)"
speedup-symlink-lookup/python.exe -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/a' * 2"   "os.path.realpath(path)"
echo 20 dirs
main/python.exe                   -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/a' * 20"  "os.path.realpath(path)"
speedup-symlink-lookup/python.exe -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/a' * 20"  "os.path.realpath(path)"
echo 200 dirs
main/python.exe                   -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/a' * 200" "os.path.realpath(path)"
speedup-symlink-lookup/python.exe -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/a' * 200" "os.path.realpath(path)"

echo ---

ln -nsfw . tmp/b
echo 2 links
main/python.exe                   -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 2"      "os.path.realpath(path)"
speedup-symlink-lookup/python.exe -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 2"      "os.path.realpath(path)"
echo 20 links
main/python.exe                   -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 20"     "os.path.realpath(path)"
speedup-symlink-lookup/python.exe -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 20"     "os.path.realpath(path)"
echo 200 links
main/python.exe                   -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 200"    "os.path.realpath(path)"
speedup-symlink-lookup/python.exe -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 200"    "os.path.realpath(path)"
echo 2,000 links
main/python.exe                   -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 2_000"  "os.path.realpath(path)"
speedup-symlink-lookup/python.exe -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 2_000"  "os.path.realpath(path)"
echo 20,000 links
main/python.exe                   -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 20_000" "os.path.realpath(path)"
speedup-symlink-lookup/python.exe -m timeit -s "import os; path = os.getcwd() + '/tmp' + '/b' * 20_000" "os.path.realpath(path)"
```

</details>

<details><summary>2 dirs - no difference</summary>

```none
10000 loops, best of 5: 35.2 usec per loop
10000 loops, best of 5: 35.2 usec per loop
```

</details>

<details><summary>20 dirs - no difference</summary>

```none
2000 loops, best of 5: 129 usec per loop
2000 loops, best of 5: 129 usec per loop
```

</details>

<details><summary>200 dirs - no difference</summary>

```none
100 loops, best of 5: 2.4 msec per loop
100 loops, best of 5: 2.39 msec per loop
```

</details>

---

<details><summary>2 links - 1.18x faster</summary>

```none
5000 loops, best of 5: 41 usec per loop
10000 loops, best of 5: 34.7 usec per loop
```

</details>

<details><summary>20 links - 3.31x faster</summary>

```none
2000 loops, best of 5: 131 usec per loop
10000 loops, best of 5: 39.6 usec per loop
```

</details>

<details><summary>200 links - 12.03x faster</summary>

```none
200 loops, best of 5: 1.05 msec per loop
5000 loops, best of 5: 87.3 usec per loop
```

</details>

<details><summary>2,000 links - 18.33x faster</summary>

```none
50 loops, best of 5: 9.77 msec per loop
500 loops, best of 5: 533 usec per loop
```

</details>

<details><summary>20,000 links - 19.27x faster</summary>

```none
5 loops, best of 5: 95.4 msec per loop
50 loops, best of 5: 4.95 msec per loop
```

</details>

<!-- gh-issue-number: gh-120145 -->
* Issue: gh-120145
<!-- /gh-issue-number -->