@picnixz 2024-06-06
Just wondering, but how many symlinks do you need until you see a noticable effect? (in your case it's 200_000 but I honestly don't think you'll ever have that use case or maybe I'm biased?)