@nineteendo 2024-06-06
> how many symlinks do you need until you see a noticable effect?

2 is enough for an 11% speedup.

> I honestly don't think you'll ever have that use case

This is partially a security issue, given that paths can get really long using symlinks.