@picnixz 2024-06-06
> 2 is enough for an 11% speedup.

IMO, this is enough to motivate your feature (it's not uncommon to have a chain `symlink1->symlink2->file` I think).