@nineteendo 2024-06-06
> it's not uncommon to have a chain `symlink1->symlink2->file` I think

Note that this only works for the same symlink. e.g. `symlink->.` and `symlink1/symlink1`.
But as we're already caching them, speedup up retrieval only makes sense, right?