@barneygale 2024-06-06
How much slower is this for paths that _don't_ involve symlinks?