@nineteendo 2024-06-07
Only 3% for 20 directories. <2% for other cases.