@barneygale 2024-06-07
Then IMHO it's not worth it - symlinks are much rarer than non-symlinks, and so even a 2-3% slowdown has an outsized effect.