@nineteendo 2024-06-07
Let me benchmark again with `seen and newpath in seen`, that's a bit faster.