@nineteendo 2024-06-07
I can't measure a slowdown anymore. For 20 directories it's faster now, but that's within the margin of error.