@nineteendo 2024-06-09
Barney, is it fine now or should I close this? The overhead is 8.2 nanoseconds, on something measured in microseconds:

```zsh
wannes@Stefans-iMac ~ % python -m timeit -s "seen = {}; key = 'foo'"
50000000 loops, best of 5: 6.5 nsec per loop
wannes@Stefans-iMac ~ % python -m timeit -s "seen = {}; key = 'foo'" "seen and key in seen"
20000000 loops, best of 5: 14.7 nsec per loop
```