@nineteendo 2024-06-11
@barneygale, I moved the `is None` check inside `if not name or name == curdir`, that means the performance decrease is only 3 nanoseconds now:

```none
wannes@Stefans-iMac ~ % python -m timeit -s "seen = {}; key = 'foo'" "seen is None"        
20000000 loops, best of 5: 11.7 nsec per loop
wannes@Stefans-iMac ~ % python -m timeit -s "seen = {}; key = 'foo'" "seen and key in seen"
20000000 loops, best of 5: 14.7 nsec per loop
```